[resharding] Generate intervals for non account_id prefix trie keys
