+++ core/store/src/trie/ops/mod.rs
@@ -2,3 +2,5 @@ pub mod insert_delete;
 pub mod interface;
 pub mod resharding;
 pub mod squash;
+#[cfg(test)]
+mod tests;
